+++ eth_hash/auto.py
@@ -1,8 +1,8 @@
-from .backends import (
-    pycryptodome,
-)
-from .main import (
+from eth_hash.main import (
     Keccak256,
 )
+from eth_hash.utils import (
+    auto_choose_backend,
+)
 
-keccak = Keccak256(pycryptodome)
+keccak = Keccak256(auto_choose_backend())